rename dropboxProvider and move to providers directory
